refactor(publish): share the publish constants instead of policing two copies
A published book's bundle cannot import the @adt/types barrel — it is built
from zod schemas, and a value import puts zod in every book. The reader
therefore restated thirteen constants by hand, with two drift tests keeping
the copies honest.

Those tests caught nothing: the copies were always edited together, and they
still let COMMENTER_PIN_MAX_LENGTH sit at 6 on one side and 12 on the other.

Adds `@adt/types/limits`, a zero-import module beside the existing ./color and
./fingerprint subpaths, and re-exports it from the schema modules so the barrel
is unchanged for the worker, the API and the Studio. The reader can now import
the real values, which is what lets the mirror go.

Also in the worker: `emptyPublicationStore` moves to testing.ts, since exactly
one test used it and it was a third implementation of a 30-method interface
living in production source; drops the unused `isHtmlRequest`; and demotes six
exports that nothing outside their own module ever read.
