+++ apps/publish-service/src/app.test.ts
@@ -2,7 +2,8 @@ import { describe, expect, it } from "vitest"
 import { zipSync } from "fflate"
 import { PUBLISH_WORKER_VERSION, type Publication } from "@adt/types"
 import { createApp } from "./app.js"
-import { emptyPublicationStore, type PublicationStore } from "./store.js"
+import { type PublicationStore } from "./store.js"
+import { emptyPublicationStore } from "./testing.js"
 import { createMemoryPublicationStore, createMemoryR2Bucket } from "./testing.js"
 
 const TOKEN = "aBcDeFgHiJkLmNoPqRsTuVwXyZ012345"
