+++ apps/publish-service/src/store.ts
@@ -197,111 +197,3 @@ export interface PublicationStore {
     resolvedAt: string | null,
   ): Promise<PublishComment | null>
 }
-
-export const emptyPublicationStore: PublicationStore = {
-  async countAccessFailures() {
-    return { byClient: 0, byToken: 0 }
-  },
-  async recordAccessFailure() {},
-  async clearAccessFailures() {},
-  async findByToken() {
-    return null
-  },
-  async findRecord() {
-    return null
-  },
-  async listPublications() {
-    return []
-  },
-  async listVersions() {
-    return []
-  },
-  async findVersion() {
-    return null
-  },
-  async create(input) {
-    return {
-      version: input.publication.current_version,
-      page_manifest: input.pageManifest,
-      created_at: input.publication.created_at,
-    }
-  },
-  async addVersion() {
-    return null
-  },
-  async revoke() {
-    return null
-  },
-  async reinstate() {
-    return null
-  },
-  async deletePublication() {
-    return null
-  },
-  async setExpiry() {
-    return null
-  },
-  async setAccessCode() {
-    return null
-  },
-  async createSession(input) {
-    return { id: input.id, name: input.name, color: input.color, is_author: input.isAuthor }
-  },
-  async ensureAuthorSession(input) {
-    return { id: input.id, name: input.name, color: input.color, is_author: true }
-  },
-  async findAuthorSession() {
-    return null
-  },
-  async findSession() {
-    return null
-  },
-  async listCommenterSessions() {
-    return []
-  },
-  async listReaders() {
-    return []
-  },
-  async renameSession() {
-    return null
-  },
-  async setSessionPin() {
-    return null
-  },
-  async countCommenterSessions() {
-    return 0
-  },
-  async createComment(input) {
-    return {
-      id: input.id,
-      token: input.token,
-      version: input.version,
-      page_section_id: input.pageSectionId,
-      parent_id: input.parentId,
-      session_id: input.sessionId,
-      author_name: "Author",
-      author_color: "#8d8d8d",
-      body: input.body,
-      anchor: input.anchor,
-      resolved_at: null,
-      edited_at: null,
-      deleted_at: null,
-      created_at: input.createdAt,
-    }
-  },
-  async findComment() {
-    return null
-  },
-  async listComments() {
-    return []
-  },
-  async updateComment() {
-    return null
-  },
-  async softDeleteComment() {
-    return null
-  },
-  async setCommentResolved() {
-    return null
-  },
-}
